validation: establish a pattern for metatests (#687)
* move file

* add superchain-version_test.go

* validation: unmarshal standard release into exported struct member

and panic if we get an empty string
